+++ src/main/scala/hatdex/hat/api/models/ApiBundleContext.scala
@@ -2,40 +2,6 @@ package hatdex.hat.api.models
 
 import org.joda.time.LocalDateTime
 
-case class ApiBundlePropertySliceCondition(
-    id: Option[Int],
-    table: ApiBundlePropertySlice,          // FIXME: linking the wrong way round as discussed on slack
-    operator: ComparisonOperator,
-    value: String)
-
-object ApiBundlePropertySliceCondition {
-  def fromBundlePropertySliceCondition(condition: BundlePropertyliceconditionRow)
-                                      (field: ApiDataField) : ApiBundlePropertySliceCondition = {   // FIXME: ApiDataField?
-    new ApiBundlePropertySliceCondition(Some(condition.id),
-      table, condition.operator, condition.value,
-      ComparisonOperators.fromString(condition.operator))
-  }
-}
-
-case class ApiBundlePropertySlice(
-    id: Option[Int],
-    dateCreated: Option[LocalDateTime],
-    lastUpdated: Option[LocalDateTime],
-    table: ApiDataTable,                      // FIXME: working with properties, not tables here?
-    conditions: Seq[ApiBundleTableCondition]) // FIXME: ApiBundleTableCondition -> ApiBundlePropertySliceCondition
-
-object ApiBundlePropertySlice {
-  // FIXME: working with properties, not tables here?
-  // FIXME: returning ApiBundleTableSlice, should be returning ApiBundlePropertySlice
-  def fromBundleTableSlice(slice: BundleTablesliceRow)(table: ApiDataTable) : ApiBundleTableSlice = {
-    new ApiBundleTableSlice(Some(slice.id),
-      Some(slice.dateCreated), Some(slice.lastUpdated),
-      table, Seq())
-  }
-}
-
-case class ApiSelection
-
 case class ApiEntitySelection(
     id: Option[Int],
     dateCreated: Option[LocalDateTime],
